implement received damage animation for player and enemies; change all float properties in scriptable objects to be of type double; implement method to flip enemy canvas

+++ Game/Assets/Data/Elements/Element.cs
@@ -8,12 +8,12 @@ public class Element : ScriptableObject
     [SerializeField] private new string name;
     [SerializeField] private Element weakerElement;
     [SerializeField] private Element strongerElement;
-    [SerializeField] private float damageIncreaseModifier = 1.2f;
-    [SerializeField] private float damageDecreaseModifier = 0.8f;
+    [SerializeField] private double damageIncreaseModifier = 1.2f;
+    [SerializeField] private double damageDecreaseModifier = 0.8f;
 
     public string Name { get => name; set => name = value; }
     public Element WeakerElement { get => weakerElement; set => weakerElement = value; }
     public Element StrongerElement { get => strongerElement; set => strongerElement = value; }
-    public float DamageIncreaseModifier { get => damageIncreaseModifier; set => damageIncreaseModifier = value; }
-    public float DamageDecreaseModifier { get => damageDecreaseModifier; set => damageDecreaseModifier = value; }
+    public double DamageIncreaseModifier { get => damageIncreaseModifier; set => damageIncreaseModifier = value; }
+    public double DamageDecreaseModifier { get => damageDecreaseModifier; set => damageDecreaseModifier = value; }
 }
